pyproject.toml
[tool.poetry]
name = "sdlc-agents"
version = "0.1.0"
description = "Multi-agent system for automated SDLC with Azure DevOps integration"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "sdlc_agents", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
# LLM providers
openai = "^1.0.0"
ollama = "^0.1.0"
# ADO integration
azure-devops = "7.1.0b4"
# Database
clickhouse-connect = "^0.7.0"
# Git operations
gitpython = "^3.1.0"
# CLI and utilities
click = "^8.1.0"
rich = "^13.0.0"
pydantic = "^2.0.0"
pydantic-settings = "^2.0.0"
python-dotenv = "^1.0.0"
aiohttp = "^3.9.0"
asyncio = "^3.4.3"
pyyaml = "^6.0"
jinja2 = "^3.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
black = "^23.0.0"
ruff = "^0.1.0"
mypy = "^1.7.0"

[tool.poetry.scripts]
sdlc-agent = "sdlc_agents.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
